+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Possible sections are:
 
 - Add an empty line between each book, makes output more legible
 - Check that the last modified time of the images is `2000-01-01 00:00:01`
+- Check against the presence of EXIF metadata
 
 ### Changed